Small diff

+++ BalanceBoard.py
@@ -37,8 +37,8 @@ class DistanceLine:
     def compute_line_end_xy(self):
         sin = np.sin(self.angle)
         cos = np.cos(self.angle)
-        end_x = self.start_xy[0] + cos * self.distance
-        end_y = self.start_xy[1] - sin * self.distance
+        end_x = int(self.start_xy[0] + cos * self.distance)
+        end_y = int(self.start_xy[1] - sin * self.distance)
         return end_x, end_y
 
     def _compute_polygon_points(self, x_center, y_center):
